# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Emerald Support Bot API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Article {
  id: number;
  title: string;
  description: string;
  body: string;
  category: string;
  sourceUrl: string;
  trustScore: number;
  lastUpdated: string;
}

export interface ChatRequest {
  message: string;
  sessionId: string;
}

export interface SourceReference {
  title: string;
  url: string;
  type: string;
}

export interface ChatResponse {
  reply: string;
  intent: string;
  trustScore: number;
  isSecurityAlert: boolean;
  isFinancialAdvice: boolean;
  sources: SourceReference[];
  ciBreakdown: string;
  lastUpdated: string;
  relatedArticles?: Article[];
}

export interface ChatMessage {
  role: string;
  content: string;
  timestamp: string;
}

export interface EscalateRequest {
  sessionId: string;
  subject: string;
  chatHistory: ChatMessage[];
}

export type EscalateResponseTicketComment = {
  body: string;
};

export type EscalateResponseTicketCustomField = {
  id: string;
  value: string;
};

export type EscalateResponseTicket = {
  subject: string;
  comment: EscalateResponseTicketComment;
  priority: string;
  custom_fields: EscalateResponseTicketCustomField[];
};

export interface EscalateResponse {
  success: boolean;
  ticket: EscalateResponseTicket;
}

export type ListArticlesParams = {
  category?: string;
  /**
   * Search query to filter articles by title/description
   */
  q?: string;
};