# lib/api-zod/src/generated/types/escalateResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Emerald Support Bot API
 * OpenAPI spec version: 0.1.0
 */
import type { EscalateResponseTicket } from "./escalateResponseTicketPayload";

export interface EscalateResponse {
  success: boolean;
  ticket: EscalateResponseTicket;
}